medical_ui/backend/requirements.txt
# Django Auth + Roles MVP
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

